Add Release Drafter configuration file
This commit includes the addition of a 'release-drafter.yml' file, which is used to automatically generate a summary of changes for each new release. As a part of GitHub Actions workflows updates, it streamlines the release process and enhances automation by keeping track of the changes in each merged pull request.