Refactor and extend Linode API usage in linode4j
Changed the LinodeInterface package location and added new classes and methods related to the Linode instance management. The changes also include a new LinodeController and the respective request and response classes for creating an instance. Additionally, created classes to model the instance type and related types such as Price and Backups, as well as a class for the description of regions. The LinodeService was also updated with caching and instance creation capabilities.

The reason for these changes was to enhance the program's usage of the Linode API, by expanding program capabilities to list instance types, list regions, and create a Linode instance. This provides a more comprehensive interaction with the Linode API.